Support multi-server - breaks scheduled tasks
We are now able to support multiple servers, though the
scheduled tasks are completely broken as we have no way
to match guilds with configurations at the moment. We
need to store guild configurations in the database
instead of in the config file. The config file should
then only include configuration to run the bot itself
and all guild-specific, or campaign-specific, config
will be done through bot commands.