Channel and author check/Inventory sub-commands
Added a channel and author check to resolve an earlier issue. Added inventory sub-commands.

Bot will now check if the message is in the channel as the one it was originally called from and also checks that the author is the same as the one who made the original command call.

New inventory commands are:
!inv
!inv add
!inv remove