Update to Channel and author check/Inventory subcommands
This is an update to the previous commit with the suggested changes from jwhett implemented.

Added new import for TimeourError
Adjusted try catch
Added a channel and author check to resolve an earlier issue.
Added inventory subcommands.

Bot will now check if the message is in the channel as the one it was originally called from and also checks that the author is the same as the one who made the original command call.

New inventory commands are:
!inv
!inv add
!inv remove