Migrate to discord.py v2 (#15)
<!--- Provide a general summary of your changes in the Title above -->

## Description
Discord.py 2.0 [made a few breaking
changes](https://discordpy.readthedocs.io/en/stable/migrating.html#).
This is also coupled with the fact that Python 3.10 changed a couple of
things about how asyncio works.

In order to keep the bot running with this new version, we need to put
the alert dispatcher inside the `on_ready` async function. I also wrap
the bot run in a try/finally block just so I know when its dying

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Once all deps are upgraded, the bot no longer works with discord.py 2.0

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
I tested this locally. If it actually connects to discord, then its
working

## Screenshots (if appropriate):
![Screenshot from 2022-10-16
23-59-55](https://user-images.githubusercontent.com/4042877/196086488-df19bd62-921f-4f9b-9b9e-a78183eba4eb.png)
![Screenshot from 2022-10-17
00-01-10](https://user-images.githubusercontent.com/4042877/196087049-26194149-1f5e-4883-a5e9-2a923e6ce475.png)


## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.